+++ NEWS.md
@@ -1,6 +1,7 @@
 # arkhe 1.7.0.9000
 ## New classes and methods
-* Add `assert_nrow` and `assert_ncol` to check the number of rows/columns.
+* Add `append_column()` to add a named vector as a column in a `data.frame`.
+* Add `assert_nrow()` and `assert_ncol()` to check the number of rows/columns.
 
 ## Enhancements
 * `seek_rows()` and `seek_columns()` gained a new `names` argument.